fix: sherlock-audit/2023-03-notional-judging#12 incorrect toUint72

+++ contracts/math/SafeUint256.sol
@@ -128,7 +128,7 @@ library SafeUint256 {
     }
 
     function toUint72(uint256 x) internal pure returns (uint72) {
-        require(x <= type(uint8).max);
+        require(x <= type(uint72).max);
         return uint72(x);
     }
